+++ agrest-cayenne/src/main/java/io/agrest/cayenne/processor/delete/CayenneDeleteStartStage.java
@@ -1,6 +1,9 @@
 package io.agrest.cayenne.processor.delete;
 
 import io.agrest.cayenne.persister.ICayennePersister;
+import io.agrest.meta.AgDataMap;
+import io.agrest.meta.AgEntity;
+import io.agrest.meta.AgEntityOverlay;
 import io.agrest.processor.ProcessingContext;
 import io.agrest.processor.Processor;
 import io.agrest.processor.ProcessorOutcome;
@@ -13,26 +16,41 @@ import org.apache.cayenne.di.Inject;
  */
 public class CayenneDeleteStartStage implements Processor<DeleteContext<?>> {
 
-    private static final String DELETE_OBJECT_CONTEXT_ATTRIBITE = "deleteContext";
+    private static final String DELETE_OBJECT_CONTEXT_ATTRIBUTE = "deleteContext";
 
     /**
      * Returns Cayenne ObjectContext previously stored in the ProcessingContext
      * by this stage.
      */
     public static ObjectContext cayenneContext(ProcessingContext<?> context) {
-        return (ObjectContext) context.getAttribute(CayenneDeleteStartStage.DELETE_OBJECT_CONTEXT_ATTRIBITE);
+        return (ObjectContext) context.getAttribute(CayenneDeleteStartStage.DELETE_OBJECT_CONTEXT_ATTRIBUTE);
     }
 
-    private ICayennePersister persister;
+    private final ICayennePersister persister;
+    private final AgDataMap dataMap;
 
-    public CayenneDeleteStartStage(@Inject ICayennePersister persister) {
+    public CayenneDeleteStartStage(@Inject ICayennePersister persister, @Inject AgDataMap dataMap) {
         this.persister = persister;
+        this.dataMap = dataMap;
     }
 
     @Override
     public ProcessorOutcome execute(DeleteContext<?> context) {
-        context.setAttribute(DELETE_OBJECT_CONTEXT_ATTRIBITE, persister.newContext());
+        initAgEntity(context);
+        initCayenneContext(context);
         return ProcessorOutcome.CONTINUE;
+    }
+
+    protected <T> void initAgEntity(DeleteContext<T> context) {
+        AgEntityOverlay<T> overlay = context.getEntityOverlay();
+        AgEntity<T> entity = dataMap.getEntity(context.getType());
+        context.setAgEntity(
+                overlay != null ? overlay.resolve(dataMap, entity) : entity
+        );
+    }
+
+    protected <T> void initCayenneContext(DeleteContext<T> context) {
+        context.setAttribute(DELETE_OBJECT_CONTEXT_ATTRIBUTE, persister.newContext());
     }
 }