+++ agrest-engine/src/main/java/io/agrest/DeleteBuilder.java
@@ -1,5 +1,8 @@
 package io.agrest;
 
+import io.agrest.access.DeleteAuthorizer;
+import io.agrest.meta.AgEntityOverlay;
+
 import java.util.Map;
 
 /**
@@ -26,6 +29,22 @@ public interface DeleteBuilder<T> {
      * @since 1.20
      */
     DeleteBuilder<T> parent(Class<?> parentType, Map<String, Object> parentIds, String relationshipFromParent);
+
+    /**
+     * Installs request-scoped {@link AgEntityOverlay} that allows to customize, add or redefine request entity structure
+     * This method can be called multiple times to add more than one overlay.
+     *
+     * @param overlay overlay descriptor
+     * @return this builder instance
+     * @since 4.8
+     */
+    DeleteBuilder<T> entityOverlay(AgEntityOverlay<T> overlay);
+
+    /**
+     * @return this builder instance
+     * @since 4.8
+     */
+    DeleteBuilder<T> authorizer(DeleteAuthorizer<T> authorizer);
 
     SimpleResponse delete();
 }